# Conflicts:
#	src/main/kotlin/fr/motrelou/bot/Utils.kt
#	src/main/kotlin/fr/motrelou/bot/api/API.kt
#	src/main/kotlin/fr/motrelou/bot/main.kt

+++ src/main/kotlin/fr/motrelou/bot/commands/GetAll.kt
@@ -10,6 +10,7 @@ import fr.motrelou.bot.api.API
 import fr.motrelou.bot.embeds.embedServerError
 import fr.motrelou.bot.embeds.listEmbed
 import fr.motrelou.bot.excpetions.APIException
+import fr.motrelou.bot.getNextPage
 import fr.motrelou.bot.kord
 
 class GetAll(private val api: API): RegisterCommand(Command(
@@ -33,29 +34,25 @@ class GetAll(private val api: API): RegisterCommand(Command(
 }){
 	init {
 		kord.on<ButtonInteractionCreateEvent> {
-			val resp = interaction.deferPublicResponse()
-			val page = if (interaction.component.customId!!.startsWith("next")) {
-				interaction.component.customId!!.substring(4).toLong() + 1
-			} else if(interaction.component.customId!!.startsWith("prev")){
-				interaction.component.customId!!.substring(4).toLong() - 1
-			}else{
-				return@on
-			}
-			val mots = api.mot(page)
-			if (mots.isNotEmpty()) {
-				interaction.message.edit {
-					listEmbed(mots, interaction.message.getGuild().id, page)
-				}
-			} else {
-				interaction.message.edit {
-					actionRow {
-						interactionButton(ButtonStyle.Primary, "prev${page-1}") {
-							label = "⬅️ Prev"
+			if(interaction.message.data.interaction.value?.name == "getall") {
+				val resp = interaction.deferPublicResponse()
+				val page = interaction.component.getNextPage()!!
+				val mots = api.mot(page)
+				if(mots.isNotEmpty()) {
+					interaction.message.edit {
+						listEmbed(mots, interaction.message.getGuild().id, page)
+					}
+				} else {
+					interaction.message.edit {
+						actionRow {
+							interactionButton(ButtonStyle.Primary, "prev${page - 1}") {
+								label = "⬅️ Prev"
+							}
 						}
 					}
 				}
+				resp.delete()
 			}
-			resp.delete()
 		}
 	}
 }
